Rename query_string to permalink
